refactor: move providers to modular structure with pipeline
