+++ package.json
@@ -63,15 +63,11 @@
   "description": "A coding agent to enhance your development workflow.",
   "devDependencies": {
     "@ai-sdk/anthropic": "^3.0.29",
-    "@ai-sdk/cerebras": "^2.0.24",
     "@ai-sdk/google": "^3.0.16",
-    "@ai-sdk/huggingface": "^1.0.23",
     "@ai-sdk/mcp": "^1.0.14",
     "@ai-sdk/openai": "^3.0.21",
     "@ai-sdk/openai-compatible": "^2.0.22",
     "@ai-sdk/provider": "^3.0.5",
-    "@ai-sdk/xai": "^3.0.41",
-    "@aihubmix/ai-sdk-provider": "^1.0.3",
     "@biomejs/biome": "^2.3.11",
     "@fastify/type-provider-typebox": "6.1.0",
     "@microsoft/api-extractor": "^7.55.2",
