+++ src/model.ts
@@ -1,2056 +1,37 @@
 import { createAnthropic } from '@ai-sdk/anthropic';
-import { createCerebras } from '@ai-sdk/cerebras';
-import { createGoogleGenerativeAI } from '@ai-sdk/google';
-import { createHuggingFace } from '@ai-sdk/huggingface';
 import { createOpenAI } from '@ai-sdk/openai';
 import { createOpenAICompatible } from '@ai-sdk/openai-compatible';
-import { createXai } from '@ai-sdk/xai';
-import { createAihubmix } from '@aihubmix/ai-sdk-provider';
-import { createOpenRouter } from '@openrouter/ai-sdk-provider';
 import type { LanguageModelV3 } from '@ai-sdk/provider';
-import {
-  extractReasoningMiddleware,
-  type LanguageModelMiddleware,
-  wrapLanguageModel,
-} from 'ai';
 import assert from 'assert';
 import defu from 'defu';
-import {
-  AntigravityProvider,
-  createAntigravityProvider,
-  GithubProvider,
-} from 'oauth-providers';
 import { ConfigManager, type ProviderConfig } from './config';
 import type { Context } from './context';
 import { PluginHookType } from './plugin';
 import { getThinkingConfig } from './thinking-config';
-import { rotateApiKey } from './utils/apiKeyRotation';
-import { mergeSystemMessagesMiddleware } from './utils/mergeSystemMessagesMiddleware';
-import { prependSystemMessageMiddleware } from './utils/prependSystemMessageMiddleware';
+import { models } from './modelMap';
+import {
+  ApiFormat,
+  providers,
+  createModelCreator,
+  type Provider,
+  type ProvidersMap,
+  type Model,
+  type ModelModalities,
+} from './providers';
 
-export interface ModelModalities {
-  input: ('text' | 'image' | 'audio' | 'video' | 'pdf')[];
-  output: ('text' | 'audio' | 'image')[];
-}
-
-interface ModelCost {
-  input: number;
-  output: number;
-  cache_read?: number;
-  cache_write?: number;
-}
-
-interface ModelLimit {
-  context: number;
-  output: number;
-}
-
-export interface Model {
-  id: string;
-  name: string;
-  shortName?: string;
-  attachment: boolean;
-  reasoning: boolean;
-  temperature: boolean;
-  tool_call: boolean;
-  knowledge: string;
-  release_date: string;
-  last_updated: string;
-  modalities: ModelModalities;
-  open_weights: boolean;
-  cost: ModelCost;
-  limit: ModelLimit;
-}
-
-export interface Provider {
-  id: string;
-  env: string[];
-  name: string;
-  apiEnv?: string[];
-  api?: string;
-  doc: string;
-  models: Record<string, string | Omit<Model, 'id' | 'cost'>>;
-  createModel?: (
-    name: string,
-    provider: Provider,
-    options: {
-      globalConfigDir: string;
-      setGlobalConfig: (key: string, value: string, isGlobal: boolean) => void;
-    },
-  ) => Promise<LanguageModelV3> | LanguageModelV3;
-  apiFormat?: 'anthropic' | 'openai' | 'responses';
-  options?: {
-    baseURL?: string;
-    apiKey?: string;
-    headers?: Record<string, string>;
-    httpProxy?: string;
-  };
-  source?: 'built-in' | string;
-}
-
-import { createProxyFetch } from './utils/proxy';
-
-export type ProvidersMap = Record<string, Provider>;
+// Re-export types for backward compatibility
+export {
+  ApiFormat,
+  type Provider,
+  type ProvidersMap,
+  type Model,
+  type ModelModalities,
+};
 export type ModelMap = Record<string, Omit<Model, 'id' | 'cost'>>;
-
-/**
- * Inject proxy support into AI SDK configuration
- * Priority: Provider-level proxy > Global proxy
- *
- * @param config - SDK configuration object
- * @param provider - Provider configuration
- * @returns Config with proxy fetch injected if proxy is configured
- */
-function withProxyConfig<T extends Record<string, any>>(
-  config: T,
-  provider: Provider,
-): T {
-  const proxyUrl = provider.options?.httpProxy;
-
-  if (proxyUrl) {
-    const proxyFetch = createProxyFetch(proxyUrl);
-    return {
-      ...config,
-      fetch: proxyFetch,
-    };
-  }
-
-  return config;
-}
-
-export const models: ModelMap = {
-  'deepseek-v3-0324': {
-    name: 'DeepSeek-V3-0324',
-    shortName: 'DeepSeek V3',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-06',
-    release_date: '2025-03-24',
-    last_updated: '2025-03-24',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 128000, output: 8192 },
-  },
-  'deepseek-v3-1': {
-    name: 'DeepSeek V3.1',
-    shortName: 'DeepSeek V3.1',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-07',
-    release_date: '2025-08-21',
-    last_updated: '2025-08-21',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 163840, output: 163840 },
-  },
-  'deepseek-v3-1-terminus': {
-    name: 'DeepSeek V3.1 Terminus',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-07',
-    release_date: '2025-09-22',
-    last_updated: '2025-09-22',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 65536 },
-  },
-  'deepseek-v3-2-exp': {
-    name: 'DeepSeek V3.2 Exp',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-09',
-    release_date: '2025-09-29',
-    last_updated: '2025-09-29',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 65536 },
-  },
-  'deepseek-v3.2': {
-    name: 'DeepSeek V3.2',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-12',
-    release_date: '2025-12-01',
-    last_updated: '2025-12-01',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 65536 },
-  },
-  'deepseek-v3.2-speciale': {
-    name: 'DeepSeek V3.2 Speciale',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: false,
-    knowledge: '2025-12',
-    release_date: '2025-12-01',
-    last_updated: '2025-12-01',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 131072 },
-  },
-  'deepseek-r1-0528': {
-    name: 'DeepSeek-R1-0528',
-    shortName: 'DeepSeek R1',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-06',
-    release_date: '2025-05-28',
-    last_updated: '2025-05-28',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 65536, output: 8192 },
-  },
-  'doubao-seed-1.6': {
-    name: 'Doubao Seed 1.6',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-01',
-    release_date: '2025-06-11',
-    last_updated: '2025-09-23',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 163840, output: 163840 },
-  },
-  'kimi-k2': {
-    name: 'Kimi K2',
-    attachment: false,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-10',
-    release_date: '2025-07-11',
-    last_updated: '2025-07-11',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 16384 },
-  },
-  'kimi-k2-turbo-preview': {
-    name: 'Kimi K2 Turbo',
-    attachment: false,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-10',
-    release_date: '2025-07-14',
-    last_updated: '2025-07-14',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 16384 },
-  },
-  'kimi-k2-0905': {
-    name: 'Kimi K2 Instruct 0905',
-    shortName: 'Kimi K2 0905',
-    attachment: false,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-10',
-    release_date: '2025-09-05',
-    last_updated: '2025-09-05',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 262144, output: 16384 },
-  },
-  'kimi-k2-thinking': {
-    name: 'Kimi K2 Thinking',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-08',
-    release_date: '2025-11-06',
-    last_updated: '2025-11-06',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 262144, output: 262144 },
-  },
-  'kimi-k2-thinking-turbo': {
-    name: 'Kimi K2 Thinking Turbo',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-08',
-    release_date: '2025-11-06',
-    last_updated: '2025-11-06',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 262144, output: 262144 },
-  },
-  'kimi-k2.5': {
-    name: 'Kimi K2.5',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-01',
-    release_date: '2026-01-27',
-    last_updated: '2026-01-27',
-    modalities: { input: ['text', 'image', 'video'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 262144, output: 262144 },
-  },
-  'qwen3-coder-480b-a35b-instruct': {
-    name: 'Qwen3 Coder 480B A35B Instruct',
-    shortName: 'Qwen3 Coder',
-    attachment: false,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-04',
-    release_date: '2025-07-23',
-    last_updated: '2025-07-23',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 262144, output: 66536 },
-  },
-  'qwen3-coder-plus': {
-    name: 'Qwen3 Coder Plus',
-    attachment: false,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-04',
-    release_date: '2025-07-23',
-    last_updated: '2025-07-23',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 1048576, output: 65536 },
-  },
-  'qwen3-235b-a22b-07-25': {
-    name: 'Qwen3 235B A22B Instruct 2507',
-    shortName: 'Qwen3',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-04',
-    release_date: '2025-04-28',
-    last_updated: '2025-07-21',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 262144, output: 131072 },
-  },
-  'qwen3-max': {
-    name: 'Qwen3 Max',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-09',
-    release_date: '2025-09-05',
-    last_updated: '2025-09-05',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 262144, output: 32768 },
-  },
-  'gemini-2.5-flash': {
-    name: 'Gemini 2.5 Flash',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-01',
-    release_date: '2025-03-20',
-    last_updated: '2025-06-05',
-    modalities: {
-      input: ['text', 'image', 'audio', 'video', 'pdf'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 1048576, output: 65536 },
-  },
-  'gemini-2.5-flash-preview-09-2025': {
-    name: 'Gemini 2.5 Flash Preview 2025 09',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-01',
-    release_date: '2025-09-25',
-    last_updated: '2025-09-25',
-    modalities: {
-      input: ['text', 'image', 'audio', 'video', 'pdf'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 1048576, output: 65536 },
-  },
-  'gemini-2.5-flash-lite-preview-06-17': {
-    name: 'Gemini 2.5 Flash Lite Preview 06-17',
-    shortName: 'Gemini 2.5 Flash Lite',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-01',
-    release_date: '2025-06-17',
-    last_updated: '2025-06-17',
-    modalities: {
-      input: ['text', 'image', 'audio', 'video', 'pdf'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 65536, output: 65536 },
-  },
-  'gemini-2.5-pro': {
-    name: 'Gemini 2.5 Pro',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-01',
-    release_date: '2025-03-20',
-    last_updated: '2025-06-05',
-    modalities: {
-      input: ['text', 'image', 'audio', 'video', 'pdf'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 1048576, output: 65536 },
-  },
-  'gemini-3-pro-preview': {
-    name: 'Gemini 3 Pro Preview',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-01',
-    release_date: '2025-01-01',
-    last_updated: '2025-01-01',
-    modalities: {
-      input: ['text', 'image', 'audio', 'video', 'pdf'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 200000, output: 65536 },
-  },
-  'gemini-3-flash-preview': {
-    name: 'Gemini 3 Flash Preview',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-01',
-    release_date: '2025-12-17',
-    last_updated: '2025-12-17',
-    modalities: {
-      input: ['text', 'image', 'video', 'audio', 'pdf'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 1048576, output: 65536 },
-  },
-  'grok-4': {
-    name: 'Grok 4',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-07',
-    release_date: '2025-07-09',
-    last_updated: '2025-07-09',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 256000, output: 64000 },
-  },
-  'grok-code-fast-1': {
-    name: 'Grok Code Fast 1',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-08',
-    release_date: '2025-08-20',
-    last_updated: '2025-08-20',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 256000, output: 32000 },
-  },
-  'grok-4-fast': {
-    name: 'Grok 4 Fast',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-11',
-    release_date: '2025-08-19',
-    last_updated: '2025-08-19',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 2000000, output: 2000000 },
-  },
-  'grok-4.1-fast': {
-    name: 'Grok 4.1 Fast',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-10',
-    release_date: '2025-11-19',
-    last_updated: '2025-11-19',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 2000000, output: 2000000 },
-  },
-  'claude-3-5-sonnet-20241022': {
-    name: 'Claude Sonnet 3.5 v2',
-    shortName: 'Sonnet 3.5',
-    attachment: true,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-04-30',
-    release_date: '2024-10-22',
-    last_updated: '2024-10-22',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 8192 },
-  },
-  'claude-3-7-sonnet': {
-    name: 'Claude Sonnet 3.7',
-    shortName: 'Sonnet 3.7',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-10-31',
-    release_date: '2025-02-19',
-    last_updated: '2025-02-19',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 64000 },
-  },
-  'claude-4-sonnet': {
-    name: 'Claude Sonnet 4',
-    shortName: 'Sonnet 4',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-03-31',
-    release_date: '2025-05-22',
-    last_updated: '2025-05-22',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 64000 },
-  },
-  'claude-4-opus': {
-    name: 'Claude Opus 4',
-    shortName: 'Opus 4',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-03-31',
-    release_date: '2025-05-22',
-    last_updated: '2025-05-22',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 32000 },
-  },
-  'gpt-oss-120b': {
-    name: 'GPT OSS 120B',
-    shortName: 'GPT OSS',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-08',
-    release_date: '2025-08-05',
-    last_updated: '2025-08-05',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 32768 },
-  },
-  'gpt-5.1-codex': {
-    name: 'GPT-5.1-Codex',
-    attachment: false,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-09-30',
-    release_date: '2025-11-13',
-    last_updated: '2025-11-13',
-    modalities: {
-      input: ['text', 'image'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 400000, output: 128000 },
-  },
-  'gpt-5.1-codex-mini': {
-    name: 'GPT-5.1-Codex-mini',
-    attachment: false,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-09-30',
-    release_date: '2025-11-13',
-    last_updated: '2025-11-13',
-    modalities: {
-      input: ['text', 'image'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 400000, output: 100000 },
-  },
-  'gpt-5.1-codex-max': {
-    name: 'GPT 5.1 Codex Max',
-    attachment: true,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-09-30',
-    release_date: '2025-11-13',
-    last_updated: '2025-11-13',
-    modalities: {
-      input: ['text', 'image'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 400000, output: 128000 },
-  },
-  'gpt-5.1': {
-    name: 'GPT-5.1',
-    attachment: true,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-09-30',
-    release_date: '2025-11-13',
-    last_updated: '2025-11-13',
-    modalities: {
-      input: ['text', 'image'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 400000, output: 128000 },
-  },
-  'gpt-5': {
-    name: 'GPT-5',
-    attachment: true,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-09-30',
-    release_date: '2025-08-07',
-    last_updated: '2025-08-07',
-    modalities: {
-      input: ['text', 'audio', 'image', 'video'],
-      output: ['text', 'audio', 'image'],
-    },
-    open_weights: false,
-    limit: { context: 400000, output: 128000 },
-  },
-  'gpt-5-mini': {
-    name: 'GPT-5 Mini',
-    attachment: true,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-05-30',
-    release_date: '2025-08-07',
-    last_updated: '2025-08-07',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 272000, output: 128000 },
-  },
-  'gpt-5-codex': {
-    name: 'GPT-5-Codex',
-    attachment: false,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-09-30',
-    release_date: '2025-09-15',
-    last_updated: '2025-09-15',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 128000, output: 64000 },
-  },
-  'gpt-5.2': {
-    name: 'GPT-5.2',
-    attachment: true,
-    reasoning: true,
-    temperature: false,
-    knowledge: '2025-08-31',
-    tool_call: true,
-    release_date: '2025-12-11',
-    last_updated: '2025-12-11',
-    modalities: { input: ['text', 'image'], output: ['text', 'image'] },
-    open_weights: false,
-    limit: { context: 400000, output: 128000 },
-  },
-  'gpt-5.2-pro': {
-    name: 'GPT-5.2 Pro',
-    attachment: true,
-    reasoning: true,
-    temperature: false,
-    knowledge: '2025-08-31',
-    tool_call: true,
-    release_date: '2025-12-11',
-    last_updated: '2025-12-11',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 400000, output: 128000 },
-  },
-  'gpt-5.2-codex': {
-    name: 'GPT-5.2-Codex',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-08-31',
-    release_date: '2026-01-14',
-    last_updated: '2026-01-14',
-    modalities: {
-      input: ['text', 'image'],
-      output: ['text'],
-    },
-    open_weights: false,
-    limit: { context: 400000, output: 128000 },
-  },
-  'gpt-4.1': {
-    name: 'GPT-4.1',
-    attachment: true,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2024-04',
-    release_date: '2025-04-14',
-    last_updated: '2025-04-14',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 1047576, output: 32768 },
-  },
-  'gpt-4': {
-    name: 'GPT-4',
-    attachment: true,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2023-11',
-    release_date: '2023-11-06',
-    last_updated: '2024-04-09',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 8192, output: 8192 },
-  },
-  'gpt-4o': {
-    name: 'GPT-4o',
-    attachment: true,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2023-09',
-    release_date: '2024-05-13',
-    last_updated: '2024-05-13',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 128000, output: 16384 },
-  },
-  o3: {
-    name: 'o3',
-    attachment: true,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-05',
-    release_date: '2025-04-16',
-    last_updated: '2025-04-16',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 100000 },
-  },
-  'o3-pro': {
-    name: 'o3-pro',
-    attachment: true,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-05',
-    release_date: '2025-06-10',
-    last_updated: '2025-06-10',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 100000 },
-  },
-  'o3-mini': {
-    name: 'o3-mini',
-    attachment: false,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-05',
-    release_date: '2024-12-20',
-    last_updated: '2025-01-29',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 100000 },
-  },
-  'o4-mini': {
-    name: 'o4-mini',
-    attachment: true,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-05',
-    release_date: '2025-04-16',
-    last_updated: '2025-04-16',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 100000 },
-  },
-  'glm-4.5': {
-    name: 'GLM-4.5',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-04',
-    release_date: '2025-07-28',
-    last_updated: '2025-07-28',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 98304 },
-  },
-  'glm-4.5-air': {
-    name: 'GLM-4.5-Air',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-04',
-    release_date: '2025-07-28',
-    last_updated: '2025-07-28',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 98304 },
-  },
-  'glm-4.5-flash': {
-    name: 'GLM-4.5-Flash',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-04',
-    release_date: '2025-07-28',
-    last_updated: '2025-07-28',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 98304 },
-  },
-  'glm-4.5v': {
-    name: 'GLM-4.5V',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-04',
-    release_date: '2025-08-11',
-    last_updated: '2025-08-11',
-    modalities: { input: ['text', 'image', 'video'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 64000, output: 16384 },
-  },
-  'glm-4.6': {
-    name: 'GLM-4.6',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-04',
-    release_date: '2025-09-30',
-    last_updated: '2025-09-30',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 204800, output: 131072 },
-  },
-  'glm-4.6v': {
-    name: 'GLM-4.6V',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-04',
-    release_date: '2025-12-08',
-    last_updated: '2025-12-08',
-    modalities: { input: ['text', 'image', 'video'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 24000 },
-  },
-  'glm-4.7': {
-    name: 'GLM-4.7',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-04',
-    release_date: '2025-12-22',
-    last_updated: '2025-12-22',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 204800, output: 131072 },
-  },
-  'sonoma-dusk-alpha': {
-    name: 'Sonoma Dusk Alpha',
-    attachment: true,
-    reasoning: false,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-09',
-    release_date: '2024-09-05',
-    last_updated: '2024-09-05',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 2000000, output: 2000000 },
-  },
-  'sonoma-sky-alpha': {
-    name: 'Sonoma Sky Alpha',
-    attachment: true,
-    reasoning: false,
-    temperature: false,
-    tool_call: true,
-    knowledge: '2024-09',
-    release_date: '2024-09-05',
-    last_updated: '2024-09-05',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 2000000, output: 2000000 },
-  },
-  'claude-4.1-opus': {
-    name: 'Claude Opus 4.1',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-03-31',
-    release_date: '2025-08-05',
-    last_updated: '2025-08-05',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 32000 },
-  },
-  'claude-4-5-sonnet': {
-    name: 'Claude Sonnet 4.5 (Preview)',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-03-31',
-    release_date: '2025-09-29',
-    last_updated: '2025-09-29',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 32000 },
-  },
-  'claude-haiku-4-5': {
-    name: 'Claude Haiku 4.5',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-02-31',
-    release_date: '2025-10-15',
-    last_updated: '2025-10-15',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 64000 },
-  },
-  'claude-opus-4-5': {
-    name: 'Claude Opus 4.5',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-03-31',
-    release_date: '2025-11-24',
-    last_updated: '2025-11-24',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 200000, output: 64000 },
-  },
-  'ling-1t': {
-    name: 'InclusionAI Ling-1T',
-    attachment: true,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-10-09',
-    release_date: '2025-10-09',
-    last_updated: '2025-10-09',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 128000, output: 32000 },
-  },
-  'ring-1t': {
-    name: 'InclusionAI Ring-1T',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-10-14',
-    release_date: '2025-10-14',
-    last_updated: '2025-10-14',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 128000, output: 32000 },
-  },
-  'ring-flash-2.0': {
-    name: 'InclusionAI Ring-flash-2.0',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-09-09',
-    release_date: '2025-09-15',
-    last_updated: '2025-09-16',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 128000, output: 32000 },
-  },
-  'ling-flash-2.0': {
-    name: 'InclusionAI Ling-flash-2.0',
-    attachment: true,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-09-09',
-    release_date: '2025-09-15',
-    last_updated: '2025-09-16',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 128000, output: 32000 },
-  },
-  'ring-mini-2.0': {
-    name: 'InclusionAI Ring-mini-2.0',
-    attachment: true,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-09-09',
-    release_date: '2025-09-15',
-    last_updated: '2025-09-16',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 128000, output: 32000 },
-  },
-  'ling-mini-2.0': {
-    name: 'InclusionAI Ling-mini-2.0',
-    attachment: true,
-    reasoning: false,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-09-09',
-    release_date: '2025-09-15',
-    last_updated: '2025-09-16',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 128000, output: 32000 },
-  },
-  'minimax-m2': {
-    name: 'Minimax-M2',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '',
-    release_date: '2025-10-27',
-    last_updated: '2025-10-27',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 196608, output: 64000 },
-  },
-  'sherlock-dash-alpha': {
-    name: 'Sherlock Dash Alpha',
-    attachment: true,
-    reasoning: false,
-    temperature: false,
-    tool_call: true,
-    knowledge: 'unknown',
-    release_date: '2025-11-15',
-    last_updated: '2025-11-15',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 1840000, output: 64000 },
-  },
-  'sherlock-think-alpha': {
-    name: 'Sherlock Think Alpha',
-    attachment: true,
-    reasoning: true,
-    temperature: false,
-    tool_call: true,
-    knowledge: 'unknown',
-    release_date: '2025-11-15',
-    last_updated: '2025-11-15',
-    modalities: { input: ['text', 'image'], output: ['text'] },
-    open_weights: false,
-    limit: { context: 1840000, output: 64000 },
-  },
-  'mimo-v2-flash': {
-    name: 'MiMo V2 Flash',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '2025-12',
-    release_date: '2025-12-17',
-    last_updated: '2025-12-17',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 131072, output: 32768 },
-  },
-  'minimax-m2.1': {
-    name: 'MiniMax-M2.1',
-    attachment: false,
-    reasoning: true,
-    temperature: true,
-    tool_call: true,
-    knowledge: '',
-    release_date: '2025-12-23',
-    last_updated: '2025-12-23',
-    modalities: { input: ['text'], output: ['text'] },
-    open_weights: true,
-    limit: { context: 204800, output: 131072 },
-  },
-};
-
-export function getProviderBaseURL(provider: Provider) {
-  if (provider.options?.baseURL) {
-    return provider.options.baseURL;
-  }
-  let api = provider.api;
-  for (const env of provider.apiEnv || []) {
-    if (process.env[env]) {
-      api = process.env[env];
-      break;
-    }
-  }
-  return api;
-}
-
-function getProviderApiKey(provider: Provider) {
-  const apiKey = (() => {
-    if (provider.options?.apiKey) {
-      return provider.options.apiKey;
-    }
-    const envs = provider.env || [];
-    for (const env of envs) {
-      if (process.env[env]) {
-        return process.env[env];
-      }
-    }
-    return '';
-  })();
-  const key = rotateApiKey(apiKey);
-  return key;
-}
-
-export const createModelCreatorCompatible = (opts?: {
-  headers?: Record<string, string>;
-  fetch?: any;
-  middlewares?: LanguageModelMiddleware[];
-}) => {
-  return (name: string, provider: Provider): LanguageModelV3 => {
-    if (provider.id !== 'openai') {
-      assert(provider.api, `Provider ${provider.id} must have an api`);
-    }
-    const baseURL = getProviderBaseURL(provider);
-    const apiKey = getProviderApiKey(provider);
-    assert(baseURL, 'baseURL is required');
-    let model = createOpenAICompatible(
-      withProxyConfig(
-        {
-          name: provider.id,
-          baseURL,
-          apiKey,
-          headers: opts?.headers,
-          fetch: opts?.fetch,
-        },
-        provider,
-      ),
-    )(name);
-    if (opts?.middlewares) {
-      model = wrapLanguageModel({ model, middleware: opts.middlewares });
-    }
-    return model;
-  };
-};
-
-const defaultModelCreator = createModelCreatorCompatible();
-const defaultAnthropicModelCreator = (name: string, provider: Provider) => {
-  const baseURL = getProviderBaseURL(provider);
-  const apiKey = getProviderApiKey(provider);
-  const model = createAnthropic(
-    withProxyConfig({ apiKey, baseURL }, provider),
-  ).chat(name);
-  return wrapLanguageModel({
-    model,
-    middleware: [prependSystemMessageMiddleware],
-  });
-};
-
-const openaiModelCreator = (
-  name: string,
-  provider: Provider,
-): LanguageModelV3 => {
-  if (provider.id !== 'openai') {
-    assert(provider.api, `Provider ${provider.id} must have an api`);
-  }
-  const baseURL = getProviderBaseURL(provider);
-  const apiKey = getProviderApiKey(provider);
-  return createOpenAI(
-    withProxyConfig(
-      {
-        baseURL,
-        apiKey,
-      },
-      provider,
-    ),
-  ).chat(name);
-};
-
-const openaiModelResponseCreator = (
-  name: string,
-  provider: Provider,
-): LanguageModelV3 => {
-  if (provider.id !== 'openai') {
-    assert(provider.api, `Provider ${provider.id} must have an api`);
-  }
-  const baseURL = getProviderBaseURL(provider);
-  const apiKey = getProviderApiKey(provider);
-  return createOpenAI(
-    withProxyConfig(
-      {
-        baseURL,
-        apiKey,
-      },
-      provider,
-    ),
-  ).responses(name);
-};
-
-export const providers: ProvidersMap = {
-  'github-copilot': {
-    id: 'github-copilot',
-    name: 'GitHub Copilot',
-    source: 'built-in',
-    env: [],
-    apiEnv: [],
-    api: 'https://api.githubcopilot.com',
-    doc: 'https://github.com/settings/copilot/features',
-    models: {
-      'claude-opus-4': models['claude-4-opus'],
-      'grok-code-fast-1': models['grok-code-fast-1'],
-      'claude-3.5-sonnet': models['claude-3-5-sonnet-20241022'],
-      'o3-mini': models['o3-mini'],
-      'gpt-5-codex': models['gpt-5-codex'],
-      'gpt-4o': models['gpt-4o'],
-      'gpt-4.1': models['gpt-4.1'],
-      'o4-mini': models['o4-mini'],
-      'claude-opus-41': models['claude-4.1-opus'],
-      'gpt-5-mini': models['gpt-5-mini'],
-      'claude-3.7-sonnet': models['claude-3-7-sonnet'],
-      'gemini-2.5-pro': models['gemini-2.5-pro'],
-      'gemini-3-pro-preview': models['gemini-3-pro-preview'],
-      o3: models['o3'],
-      'claude-sonnet-4': models['claude-4-sonnet'],
-      'gpt-5.1-codex': models['gpt-5.1-codex'],
-      'gpt-5.1-codex-mini': models['gpt-5.1-codex-mini'],
-      'gpt-5.1': models['gpt-5.1'],
-      'gpt-5': models['gpt-5'],
-      'claude-3.7-sonnet-thought': models['claude-3-7-sonnet'],
-      'claude-sonnet-4.5': models['claude-4-5-sonnet'],
-      'claude-opus-4-5': models['claude-opus-4-5'],
-      'gpt-5.2': models['gpt-5.2'],
-    },
-    async createModel(name, provider, options) {
-      const apiKey = provider.options?.apiKey;
-      assert(
-        apiKey,
-        'Failed to get GitHub Copilot token, use /login to login first',
-      );
-      let account = JSON.parse(apiKey);
-      const githubProvider = new GithubProvider();
-      githubProvider.setState(account);
-      if (githubProvider.isTokenExpired()) {
-        await githubProvider.refresh();
-        account = githubProvider.getState();
-        provider.options = {
-          ...provider.options,
-          apiKey: JSON.stringify(account),
-        };
-        options.setGlobalConfig(
-          'provider.github-copilot.options.apiKey',
-          JSON.stringify(account),
-          true,
-        );
-      }
-      const token = account.copilot_token;
-      return createOpenAI({
-        baseURL: 'https://api.individual.githubcopilot.com',
-        headers: {
-          Authorization: `Bearer ${token}`,
-          'User-Agent': 'GitHubCopilotChat/0.26.7',
-          'Editor-Version': 'vscode/1.99.3',
-          'Editor-Plugin-Version': 'copilot-chat/0.26.7',
-          'Copilot-Integration-Id': 'vscode-chat',
-        },
-        // fix Failed: OpenAI API key is missing
-        apiKey: '',
-      }).chat(name);
-    },
-  },
-  openai: {
-    id: 'openai',
-    source: 'built-in',
-    env: ['OPENAI_API_KEY'],
-    apiEnv: ['OPENAI_API_BASE'],
-    name: 'OpenAI',
-    doc: 'https://platform.openai.com/docs/models',
-    models: {
-      'gpt-4.1': models['gpt-4.1'],
-      'gpt-4': models['gpt-4'],
-      'gpt-4o': models['gpt-4o'],
-      o3: models['o3'],
-      'o3-mini': models['o3-mini'],
-      'o4-mini': models['o4-mini'],
-      'gpt-5.1': models['gpt-5.1'],
-      'gpt-5.1-codex': models['gpt-5.1-codex'],
-      'gpt-5.1-codex-mini': models['gpt-5.1-codex-mini'],
-      'gpt-5.1-codex-max': models['gpt-5.1-codex-max'],
-      'gpt-5': models['gpt-5'],
-      'gpt-5-mini': models['gpt-5-mini'],
-      'gpt-5-codex': models['gpt-5-codex'],
-      'gpt-5.2': models['gpt-5.2'],
-      'gpt-5.2-pro': models['gpt-5.2-pro'],
-      'gpt-5.2-codex': models['gpt-5.2-codex'],
-    },
-    createModel: openaiModelCreator,
-  },
-  google: {
-    id: 'google',
-    source: 'built-in',
-    env: ['GOOGLE_API_KEY', 'GOOGLE_GENERATIVE_AI_API_KEY'],
-    apiEnv: ['GOOGLE_GENERATIVE_AI_API_BASE'],
-    name: 'Google',
-    doc: 'https://ai.google.dev/gemini-api/docs/pricing',
-    models: {
-      'gemini-2.5-flash': models['gemini-2.5-flash'],
-      'gemini-2.5-flash-preview-09-2025':
-        models['gemini-2.5-flash-preview-09-2025'],
-      'gemini-2.5-flash-lite': models['gemini-2.5-flash-lite-preview-06-17'],
-      'gemini-2.5-pro': models['gemini-2.5-pro'],
-      'gemini-3-pro-preview': models['gemini-3-pro-preview'],
-      'gemini-3-flash-preview': models['gemini-3-flash-preview'],
-    },
-    createModel(name, provider) {
-      const baseURL = getProviderBaseURL(provider);
-      const apiKey = getProviderApiKey(provider);
-      const google = createGoogleGenerativeAI({
-        apiKey,
-        baseURL,
-      });
-      return google(name);
-    },
-  },
-  deepseek: {
-    id: 'deepseek',
-    source: 'built-in',
-    env: ['DEEPSEEK_API_KEY'],
-    name: 'DeepSeek',
-    api: 'https://api.deepseek.com',
-    apiEnv: ['DEEPSEEK_API_BASE'],
-    doc: 'https://platform.deepseek.com/api-docs/pricing',
-    models: {
-      'deepseek-chat': models['deepseek-v3.2'],
-      'deepseek-reasoner': models['deepseek-r1-0528'],
-    },
-    createModel: defaultModelCreator,
-  },
-  xai: {
-    id: 'xai',
-    source: 'built-in',
-    env: ['XAI_API_KEY'],
-    apiEnv: ['XAI_BASE_URL'],
-    name: 'xAI',
-    doc: 'https://xai.com/docs/models',
-    models: {
-      'grok-4-1-fast': models['grok-4.1-fast'],
-      'grok-4-1-fast-non-reasoning': {
-        ...models['grok-4.1-fast'],
-        reasoning: false,
-      },
-      'grok-4': models['grok-4'],
-      'grok-4-fast': models['grok-4-fast'],
-      'grok-code-fast-1': models['grok-code-fast-1'],
-    },
-    createModel(name, provider) {
-      const api = getProviderBaseURL(provider);
-      const apiKey = getProviderApiKey(provider);
-      return createXai({
-        baseURL: api,
-        apiKey,
-      }).chat(name);
-    },
-  },
-  anthropic: {
-    id: 'anthropic',
-    source: 'built-in',
-    env: ['ANTHROPIC_API_KEY'],
-    apiEnv: ['ANTHROPIC_API_BASE'],
-    name: 'Anthropic',
-    doc: 'https://docs.anthropic.com/en/docs/models',
-    models: {
-      'claude-opus-4-20250514': models['claude-4-opus'],
-      'claude-opus-4-1-20250805': models['claude-4.1-opus'],
-      'claude-sonnet-4-20250514': models['claude-4-sonnet'],
-      'claude-sonnet-4-5-20250929': models['claude-4-5-sonnet'],
-      'claude-3-7-sonnet-20250219': models['claude-3-7-sonnet'],
-      'claude-3-7-sonnet-20250219-thinking': models['claude-3-7-sonnet'],
-      'claude-3-5-sonnet-20241022': models['claude-3-5-sonnet-20241022'],
-      'claude-haiku-4-5': models['claude-haiku-4-5'],
-      'claude-opus-4-5': models['claude-opus-4-5'],
-    },
-    apiFormat: 'anthropic',
-    createModel: defaultAnthropicModelCreator,
-  },
-  aihubmix: {
-    id: 'aihubmix',
-    source: 'built-in',
-    env: ['AIHUBMIX_API_KEY'],
-    name: 'AIHubMix',
-    api: 'https://aihubmix.com/v1',
-    doc: 'https://docs.aihubmix.com/',
-    models: {
-      'gemini-2.5-pro': models['gemini-2.5-pro'],
-      'gemini-2.5-flash': models['gemini-2.5-flash'],
-      'gemini-2.5-flash-lite': models['gemini-2.5-flash-lite-preview-06-17'],
-      'DeepSeek-R1': models['deepseek-r1-0528'],
-      'DeepSeek-V3': models['deepseek-v3-0324'],
-      'claude-opus-4-20250514': models['claude-4-opus'],
-      'claude-opus-4-1': models['claude-4.1-opus'],
-      'claude-sonnet-4-20250514': models['claude-4-sonnet'],
-      'claude-sonnet-4-5': models['claude-4-5-sonnet'],
-      'claude-3-7-sonnet-20250219': models['claude-3-7-sonnet'],
-      'claude-3-5-sonnet-20241022': models['claude-3-5-sonnet-20241022'],
-      'gpt-4.1': models['gpt-4.1'],
-      'gpt-4': models['gpt-4'],
-      'gpt-4o': models['gpt-4o'],
-      o3: models['o3'],
-      'o3-mini': models['o3-mini'],
-      'o4-mini': models['o4-mini'],
-      'gpt-5': models['gpt-5'],
-      'gpt-5-mini': models['gpt-5-mini'],
-      'glm-4.6': models['glm-4.6'],
-      'kimi-k2-thinking': models['kimi-k2-thinking'],
-      'kimi-k2-turbo-preview': models['kimi-k2-turbo-preview'],
-    },
-    createModel(name, provider) {
-      const apiKey = getProviderApiKey(provider);
-      return createAihubmix(
-        withProxyConfig(
-          {
-            apiKey,
-            headers: {
-              'APP-Code': 'TPQW7551',
-            },
-          },
-          provider,
-        ),
-      ).chat(name) as unknown as LanguageModelV3;
-    },
-  },
-  openrouter: {
-    id: 'openrouter',
-    source: 'built-in',
-    env: ['OPENROUTER_API_KEY', 'OPEN_ROUTER_API_KEY'],
-    name: 'OpenRouter',
-    doc: 'https://openrouter.ai/docs/models',
-    models: {
-      'anthropic/claude-3.5-sonnet': models['claude-3-5-sonnet-20241022'],
-      'anthropic/claude-3.7-sonnet': models['claude-3-7-sonnet'],
-      'anthropic/claude-sonnet-4': models['claude-4-sonnet'],
-      'anthropic/claude-sonnet-4.5': models['claude-4-5-sonnet'],
-      'anthropic/claude-haiku-4.5': models['claude-haiku-4-5'],
-      'anthropic/claude-opus-4': models['claude-4-opus'],
-      'anthropic/claude-opus-4.1': models['claude-4.1-opus'],
-      'anthropic/claude-opus-4.5': models['claude-opus-4-5'],
-      'deepseek/deepseek-r1-0528': models['deepseek-r1-0528'],
-      'deepseek/deepseek-chat-v3-0324': models['deepseek-v3-0324'],
-      'deepseek/deepseek-chat-v3.1': models['deepseek-v3-1'],
-      'deepseek/deepseek-v3.1-terminus': models['deepseek-v3-1-terminus'],
-      'deepseek/deepseek-v3.2-exp': models['deepseek-v3-2-exp'],
-      'deepseek/deepseek-v3.2': models['deepseek-v3.2'],
-      'deepseek/deepseek-v3.2-speciale': models['deepseek-v3.2-speciale'],
-      'openai/gpt-4.1': models['gpt-4.1'],
-      'openai/gpt-4': models['gpt-4'],
-      'openai/gpt-4o': models['gpt-4o'],
-      'openai/o3': models['o3'],
-      'openai/o3-pro': models['o3-pro'],
-      'openai/o3-mini': models['o3-mini'],
-      'openai/o4-mini': models['o4-mini'],
-      'openai/gpt-oss-120b': models['gpt-oss-120b'],
-      'openai/gpt-5.1-codex': models['gpt-5.1-codex'],
-      'openai/gpt-5.1-codex-mini': models['gpt-5.1-codex-mini'],
-      'openai/gpt-5.1-codex-max': models['gpt-5.1-codex-max'],
-      'openai/gpt-5.1': models['gpt-5.1'],
-      'openai/gpt-5': models['gpt-5'],
-      'openai/gpt-5-mini': models['gpt-5-mini'],
-      'openai/gpt-5-codex': models['gpt-5-codex'],
-      'openai/gpt-5.2': models['gpt-5.2'],
-      'openai/gpt-5.2-pro': models['gpt-5.2-pro'],
-      'openai/gpt-5.2-codex': models['gpt-5.2-codex'],
-      'google/gemini-3-flash-preview': models['gemini-3-flash-preview'],
-      'google/gemini-3-pro-preview': models['gemini-3-pro-preview'],
-      'moonshotai/kimi-k2': models['kimi-k2'],
-      'moonshotai/kimi-k2-0905': models['kimi-k2-0905'],
-      'moonshotai/kimi-k2-thinking': models['kimi-k2-thinking'],
-      'moonshotai/kimi-k2.5': models['kimi-k2.5'],
-      'qwen/qwen3-coder': models['qwen3-coder-480b-a35b-instruct'],
-      'qwen/qwen3-max': models['qwen3-max'],
-      'x-ai/grok-code-fast-1': models['grok-code-fast-1'],
-      'x-ai/grok-4': models['grok-4'],
-      'x-ai/grok-4-fast': models['grok-4-fast'],
-      'x-ai/grok-4.1-fast': models['grok-4.1-fast'],
-      'z-ai/glm-4.5': models['glm-4.5'],
-      'z-ai/glm-4.5v': models['glm-4.5v'],
-      'z-ai/glm-4.6': models['glm-4.6'],
-      'z-ai/glm-4.6v': models['glm-4.6v'],
-      'z-ai/glm-4.7': models['glm-4.7'],
-      'minimax/minimax-m2': models['minimax-m2'],
-    },
-    createModel(name, provider) {
-      const baseURL = getProviderBaseURL(provider);
-      const apiKey = getProviderApiKey(provider);
-      return createOpenRouter(
-        withProxyConfig(
-          {
-            apiKey,
-            baseURL,
-            headers: {
-              'X-Title': 'Neovate Code',
-              'HTTP-Referer': 'https://neovateai.dev/',
-            },
-          },
-          provider,
-        ),
-      ).chat(name) as unknown as LanguageModelV3;
-    },
-  },
-  iflow: {
-    id: 'iflow',
-    source: 'built-in',
-    env: ['IFLOW_API_KEY'],
-    name: 'iFlow',
-    api: 'https://apis.iflow.cn/v1/',
-    doc: 'https://iflow.cn/',
-    models: {
-      'qwen3-coder-plus': models['qwen3-coder-plus'],
-      'kimi-k2': models['kimi-k2'],
-      'kimi-k2-0905': models['kimi-k2-0905'],
-      'deepseek-v3': models['deepseek-v3-0324'],
-      'deepseek-v3.2': models['deepseek-v3-2-exp'],
-      'deepseek-r1': models['deepseek-r1-0528'],
-      'glm-4.6': models['glm-4.6'],
-      'glm-4.7': models['glm-4.7'],
-      'minimax-m2.1': models['minimax-m2.1'],
-      'qwen3-max': models['qwen3-max'],
-    },
-    createModel: createModelCreatorCompatible({
-      fetch: (url: string, options: any) => {
-        return fetch(url, {
-          ...options,
-          headers: {
-            ...options.headers,
-            'user-agent': 'iFlow-Cli',
-          },
-        });
-      },
-      middlewares: [
-        mergeSystemMessagesMiddleware,
-        extractReasoningMiddleware({
-          tagName: 'think',
-        }),
-      ],
-    }),
-  },
-  moonshotai: {
-    id: 'moonshotai',
-    source: 'built-in',
-    env: ['MOONSHOT_API_KEY'],
-    name: 'Moonshot',
-    api: 'https://api.moonshot.ai/v1',
-    doc: 'https://platform.moonshot.ai/docs/api/chat',
-    models: {
-      'kimi-k2-0905-preview': models['kimi-k2-0905'],
-      'kimi-k2-turbo-preview': models['kimi-k2-turbo-preview'],
-      'kimi-k2-thinking': models['kimi-k2-thinking'],
-      'kimi-k2-thinking-turbo': models['kimi-k2-thinking-turbo'],
-      'kimi-k2.5': models['kimi-k2.5'],
-    },
-    createModel: defaultModelCreator,
-  },
-  'moonshotai-cn': {
-    id: 'moonshotai-cn',
-    source: 'built-in',
-    env: ['MOONSHOT_API_KEY'],
-    name: 'MoonshotCN',
-    api: 'https://api.moonshot.cn/v1',
-    doc: 'https://platform.moonshot.cn/docs/api/chat',
-    models: {
-      'kimi-k2-0905-preview': models['kimi-k2-0905'],
-      'kimi-k2-turbo-preview': models['kimi-k2-turbo-preview'],
-      'kimi-k2-thinking': models['kimi-k2-thinking'],
-      'kimi-k2-thinking-turbo': models['kimi-k2-thinking-turbo'],
-      'kimi-k2.5': models['kimi-k2.5'],
-    },
-    createModel: defaultModelCreator,
-  },
-  groq: {
-    id: 'groq',
-    source: 'built-in',
-    env: ['GROQ_API_KEY'],
-    name: 'Groq',
-    api: 'https://api.groq.com/openai/v1',
-    doc: 'https://console.groq.com/docs/models',
-    models: {
-      'openai/gpt-oss-120b': models['gpt-oss-120b'],
-      'moonshotai/kimi-k2-instruct-0905': models['kimi-k2-0905'],
-    },
-    createModel: defaultModelCreator,
-  },
-  siliconflow: {
-    id: 'siliconflow',
-    source: 'built-in',
-    env: ['SILICONFLOW_API_KEY'],
-    name: 'SiliconFlow',
-    api: 'https://api.siliconflow.com/v1',
-    doc: 'https://docs.siliconflow.com',
-    models: {
-      'Qwen/Qwen3-235B-A22B-Instruct-2507': models['qwen3-235b-a22b-07-25'],
-      'Qwen/Qwen3-Coder-480B-A35B-Instruct':
-        models['qwen3-coder-480b-a35b-instruct'],
-      'moonshotai/Kimi-K2-Instruct-0905': models['kimi-k2-0905'],
-      'moonshotai/Kimi-K2-Instruct': models['kimi-k2'],
-      'deepseek-ai/DeepSeek-R1': models['deepseek-r1-0528'],
-      'deepseek-ai/DeepSeek-V3.1': models['deepseek-v3-1'],
-      'deepseek-ai/DeepSeek-V3': models['deepseek-v3-0324'],
-      'zai-org/GLM-4.5': models['glm-4.5'],
-    },
-    createModel: defaultModelCreator,
-  },
-  'siliconflow-cn': {
-    id: 'siliconflow-cn',
-    source: 'built-in',
-    env: ['SILICONFLOW_API_KEY'],
-    name: 'SiliconFlow CN',
-    api: 'https://api.siliconflow.cn/v1',
-    doc: 'https://docs.siliconflow.cn',
-    models: {
-      'Qwen/Qwen3-235B-A22B-Instruct-2507': models['qwen3-235b-a22b-07-25'],
-      'Qwen/Qwen3-Coder-480B-A35B-Instruct':
-        models['qwen3-coder-480b-a35b-instruct'],
-      'moonshotai/Kimi-K2-Instruct-0905': models['kimi-k2-0905'],
-      'moonshotai/Kimi-K2-Instruct': models['kimi-k2'],
-      'deepseek-ai/DeepSeek-R1': models['deepseek-r1-0528'],
-      'deepseek-ai/DeepSeek-V3.1': models['deepseek-v3-1'],
-      'deepseek-ai/DeepSeek-V3': models['deepseek-v3-0324'],
-      'zai-org/GLM-4.5': models['glm-4.5'],
-    },
-    createModel: defaultModelCreator,
-  },
-  modelscope: {
-    id: 'modelscope',
-    source: 'built-in',
-    env: ['MODELSCOPE_API_KEY'],
-    name: 'ModelScope',
-    api: 'https://api-inference.modelscope.cn/v1',
-    doc: 'https://modelscope.cn/docs/model-service/API-Inference/intro',
-    models: {
-      'Qwen/Qwen3-Coder-480B-A35B-Instruct':
-        models['qwen3-coder-480b-a35b-instruct'],
-      'Qwen/Qwen3-235B-A22B-Instruct-2507': models['qwen3-235b-a22b-07-25'],
-      'ZhipuAI/GLM-4.5': models['glm-4.5'],
-      'ZhipuAI/GLM-4.5V': models['glm-4.5v'],
-      'ZhipuAI/GLM-4.6': models['glm-4.6'],
-      'deepseek-ai/DeepSeek-V3.2': models['deepseek-v3.2'],
-      'deepseek-ai/DeepSeek-V3.2-Speciale': models['deepseek-v3.2-speciale'],
-    },
-    createModel: defaultModelCreator,
-  },
-  volcengine: {
-    id: 'volcengine',
-    source: 'built-in',
-    env: ['VOLCENGINE_API_KEY'],
-    name: 'VolcEngine',
-    api: 'https://ark.cn-beijing.volces.com/api/v3',
-    doc: 'https://www.volcengine.com/docs/82379/1330310',
-    models: {
-      'deepseek-v3-1-250821': models['deepseek-v3-1'],
-      'deepseek-v3-1-terminus': models['deepseek-v3-1-terminus'],
-      'doubao-seed-1-6-250615': models['doubao-seed-1.6'],
-      'kimi-k2-250905': models['kimi-k2-0905'],
-    },
-    createModel: defaultModelCreator,
-  },
-  'zai-coding-plan': {
-    id: 'zai-coding-plan',
-    source: 'built-in',
-    env: ['ZHIPU_API_KEY'],
-    name: 'Z.AI Coding Plan',
-    api: 'https://api.z.ai/api/coding/paas/v4',
-    doc: 'https://docs.z.ai/devpack/overview',
-    models: {
-      'glm-4.5-flash': models['glm-4.5-flash'],
-      'glm-4.5': models['glm-4.5'],
-      'glm-4.5-air': models['glm-4.5-air'],
-      'glm-4.5v': models['glm-4.5v'],
-      'glm-4.6': models['glm-4.6'],
-      'glm-4.6v': models['glm-4.6v'],
-      'glm-4.7': models['glm-4.7'],
-    },
-    createModel: defaultModelCreator,
-  },
-  'zhipuai-coding-plan': {
-    id: 'zhipuai-coding-plan',
-    source: 'built-in',
-    env: ['ZHIPU_API_KEY'],
-    name: 'Zhipu AI Coding Plan',
-    api: 'https://open.bigmodel.cn/api/coding/paas/v4',
-    doc: 'https://docs.bigmodel.cn/cn/coding-plan/overview',
-    models: {
-      'glm-4.6': models['glm-4.6'],
-      'glm-4.5v': models['glm-4.5v'],
-      'glm-4.5-air': models['glm-4.5-air'],
-      'glm-4.5': models['glm-4.5'],
-      'glm-4.5-flash': models['glm-4.5-flash'],
-      'glm-4.6v': models['glm-4.6v'],
-      'glm-4.7': models['glm-4.7'],
-    },
-    createModel: defaultModelCreator,
-  },
-  zhipuai: {
-    id: 'zhipuai',
-    source: 'built-in',
-    env: ['ZHIPU_API_KEY'],
-    name: 'Zhipu AI',
-    api: 'https://open.bigmodel.cn/api/paas/v4',
-    doc: 'https://docs.z.ai/guides/overview/pricing',
-    models: {
-      'glm-4.6': models['glm-4.6'],
-      'glm-4.5v': models['glm-4.5v'],
-      'glm-4.5-air': models['glm-4.5-air'],
-      'glm-4.5': models['glm-4.5'],
-      'glm-4.5-flash': models['glm-4.5-flash'],
-      'glm-4.6v': models['glm-4.6v'],
-      'glm-4.7': models['glm-4.7'],
-    },
-    createModel: defaultModelCreator,
-  },
-  zenmux: {
-    id: 'zenmux',
-    source: 'built-in',
-    env: ['ZENMUX_API_KEY'],
-    name: 'ZenMux',
-    api: 'https://zenmux.ai/api/v1',
-    doc: 'https://docs.zenmux.ai/',
-    models: {
-      'inclusionai/ling-1t': models['ling-1t'],
-      'inclusionai/ring-1t': models['ring-1t'],
-      'inclusionai/ring-flash-2.0': models['ring-flash-2.0'],
-      'inclusionai/ling-flash-2.0': models['ling-flash-2.0'],
-      'inclusionai/ring-mini-2.0': models['ring-mini-2.0'],
-      'inclusionai/ling-mini-2.0': models['ling-mini-2.0'],
-      'google/gemini-3-flash-preview': models['gemini-3-flash-preview'],
-      'google/gemini-3-pro-preview': models['gemini-3-pro-preview'],
-      'openai/gpt-5.1': models['gpt-5.1'],
-      'openai/gpt-5.1-codex': models['gpt-5.1-codex'],
-      'openai/gpt-5.1-codex-mini': models['gpt-5.1-codex-mini'],
-      'openai/gpt-5.2': models['gpt-5.2'],
-      'openai/gpt-5.2-pro': models['gpt-5.2-pro'],
-      'anthropic/claude-sonnet-4.5': models['claude-4-5-sonnet'],
-      'anthropic/claude-opus-4.1': models['claude-4.1-opus'],
-      'anthropic/claude-opus-4.5': models['claude-opus-4-5'],
-      'z-ai/glm-4.6': models['glm-4.6'],
-      'z-ai/glm-4.6v': models['glm-4.6v'],
-      'z-ai/glm-4.6v-flash': models['glm-4.6v'],
-      'deepseek/deepseek-v3.2-speciale': models['deepseek-v3.2-speciale'],
-      'deepseek/deepseek-chat': models['deepseek-v3-2-exp'],
-      'deepseek/deepseek-reasoner': models['deepseek-r1-0528'],
-      'moonshotai/kimi-k2.5': models['kimi-k2.5'],
-    },
-    createModel: createModelCreatorCompatible({
-      headers: {
-        'X-Title': 'Neovate Code',
-        'HTTP-Referer': 'https://neovateai.dev/',
-      },
-    }),
-  },
-  minimax: {
-    id: 'minimax',
-    source: 'built-in',
-    env: ['MINIMAX_API_KEY'],
-    name: 'Minimax',
-    api: 'https://api.minimaxi.io/anthropic/v1',
-    doc: 'https://platform.minimaxi.io/docs/guides/quickstart',
-    models: {
-      'minimax-m2': models['minimax-m2'],
-      'minimax-m2.1': models['minimax-m2.1'],
-    },
-    createModel(name, provider) {
-      const baseURL = getProviderBaseURL(provider);
-      const apiKey = getProviderApiKey(provider);
-      return createAnthropic(
-        withProxyConfig({ baseURL, apiKey }, provider),
-      ).chat(name);
-    },
-  },
-  'minimax-cn': {
-    id: 'minimax-cn',
-    source: 'built-in',
-    env: ['MINIMAX_API_KEY'],
-    name: 'Minimax CN',
-    api: 'https://api.minimaxi.com/anthropic/v1',
-    doc: 'https://platform.minimaxi.com/docs/guides/quickstart',
-    models: {
-      'minimax-m2': models['minimax-m2'],
-      'minimax-m2.1': models['minimax-m2.1'],
-    },
-    createModel(name, provider) {
-      const baseURL = getProviderBaseURL(provider);
-      const apiKey = getProviderApiKey(provider);
-      return createAnthropic(
-        withProxyConfig({ baseURL, apiKey }, provider),
-      ).chat(name);
-    },
-  },
-  xiaomi: {
-    id: 'xiaomi',
-    source: 'built-in',
-    env: ['MIMO_API_KEY'],
-    name: 'Xiaomi Mimo',
-    api: 'https://api.xiaomimimo.com/v1',
-    doc: 'https://platform.xiaomimimo.com/',
-    models: {
-      'mimo-v2-flash': models['mimo-v2-flash'],
-    },
-    createModel: createModelCreatorCompatible({}),
-  },
-  cerebras: {
-    id: 'cerebras',
-    source: 'built-in',
-    env: ['CEREBRAS_API_KEY'],
-    name: 'Cerebras',
-    doc: 'https://cerebras.ai/docs',
-    models: {
-      'zai-glm-4.7': {
-        ...models['glm-4.7'],
-        // ref: https://inference-docs.cerebras.ai/models/zai-glm-47
-        // default use the context of free tier
-        limit: { context: 64000, output: 40000 },
-      },
-      'gpt-oss-120b': {
-        ...models['gpt-oss-120b'],
-        // ref: https://inference-docs.cerebras.ai/models/openai-oss
-        // default use the context of free tier
-        limit: { context: 65000, output: 32000 },
-      },
-    },
-    createModel(name, provider) {
-      const apiKey = getProviderApiKey(provider);
-      return createCerebras(
-        withProxyConfig(
-          {
-            apiKey,
-            headers: {
-              'X-Cerebras-3rd-Party-Integration': 'cline',
-            },
-          },
-          provider,
-        ),
-      )(name);
-    },
-  },
-  huggingface: {
-    id: 'huggingface',
-    source: 'built-in',
-    env: ['HUGGINGFACE_API_KEY'],
-    name: 'Hugging Face',
-    doc: 'https://huggingface.co/docs/inference-providers/index',
-    models: {
-      'zai-org/GLM-4.7': models['glm-4.7'],
-      'XiaomiMiMo/MiMo-V2-Flash': models['mimo-v2-flash'],
-      'Qwen/Qwen3-Coder-480B-A35B-Instruct':
-        models['qwen3-coder-480b-a35b-instruct'],
-      'moonshotai/Kimi-K2.5': models['kimi-k2.5'],
-    },
-    createModel(name, provider) {
-      const apiKey = getProviderApiKey(provider);
-      return createHuggingFace({
-        apiKey,
-      }).languageModel(name) as unknown as LanguageModelV3;
-    },
-  },
-  poe: {
-    id: 'poe',
-    source: 'built-in',
-    env: ['POE_API_KEY'],
-    name: 'Poe',
-    api: 'https://api.poe.com/v1',
-    doc: 'https://poe.com',
-    models: {
-      'Claude-Opus-4.5': models['claude-4-opus'],
-      'Claude-Sonnet-4.5': models['claude-4-5-sonnet'],
-      'Gemini-3-Pro': models['gemini-3-pro-preview'],
-      'Gemini-2.5-Pro': models['gemini-2.5-pro'],
-      'Gemini-2.5-Flash': models['gemini-2.5-flash'],
-      'GPT-5.1': models['gpt-5.1'],
-      'GPT-5.1-Codex': models['gpt-5.1-codex'],
-      'Grok-4.1-Fast-Non-Reasoning': {
-        ...models['grok-4.1-fast'],
-        reasoning: false,
-      },
-      'Grok-4.1-Fast': models['grok-4.1-fast'],
-    },
-    createModel: defaultModelCreator,
-  },
-  antigravity: {
-    id: 'antigravity',
-    source: 'built-in',
-    env: [],
-    name: 'Antigravity',
-    doc: 'https://antigravity.google/',
-    models: {
-      'gemini-2.5-flash-lite': models['gemini-2.5-flash-lite-preview-06-17'],
-      'gemini-2.5-flash': models['gemini-2.5-flash'],
-      'gemini-2.5-flash-thinking': models['gemini-2.5-flash'],
-      'gemini-2.5-pro': models['gemini-2.5-pro'],
-      'gemini-3-flash': models['gemini-3-flash-preview'],
-      'gemini-3-pro-low': models['gemini-3-pro-preview'],
-      'gemini-3-pro-high': models['gemini-3-pro-preview'],
-      'claude-sonnet-4-5': models['claude-4-5-sonnet'],
-      'claude-sonnet-4-5-thinking': models['claude-4-5-sonnet'],
-      'claude-opus-4-5-thinking': models['claude-opus-4-5'],
-      'gpt-oss-120b-medium': models['gpt-oss-120b'],
-    },
-    async createModel(name, provider, options) {
-      const apiKey = provider.options?.apiKey;
-      assert(apiKey, 'Antigravity not logged in.');
-      let account = JSON.parse(apiKey);
-      const p = new AntigravityProvider();
-      p.setState(account);
-      if (p.isTokenExpired()) {
-        await p.refresh();
-        account = p.getState();
-        provider.options = {
-          ...provider.options,
-          apiKey: JSON.stringify(account),
-        };
-        options.setGlobalConfig(
-          'provider.antigravity.options.apiKey',
-          JSON.stringify(account),
-          true,
-        );
-      }
-      return createAntigravityProvider({
-        account,
-      })(name);
-    },
-  },
-  nvidia: {
-    id: 'nvidia',
-    source: 'built-in',
-    env: ['NVIDIA_API_KEY'],
-    name: 'NVIDIA',
-    api: 'https://integrate.api.nvidia.com/v1/',
-    doc: 'https://nvidia.com/',
-    models: {
-      'z-ai/glm4.7': models['glm-4.7'],
-      'minimaxai/minimax-m2.1': models['minimax-m2.1'],
-      'moonshotai/kimi-k2-thinking': models['kimi-k2-thinking'],
-      'moonshotai/kimi-k2.5': models['kimi-k2.5'],
-      'openai/gpt-oss-120b': models['gpt-oss-120b'],
-      'qwen/qwen3-coder-480b-a35b-instruct':
-        models['qwen3-coder-480b-a35b-instruct'],
-    },
-    createModel: createModelCreatorCompatible({
-      middlewares: [
-        extractReasoningMiddleware({
-          tagName: 'think',
-        }),
-      ],
-    }),
-  },
-  canopywave: {
-    id: 'canopywave',
-    source: 'built-in',
-    env: ['CANOPYWAVE_API_KEY'],
-    name: 'CanopyWave',
-    api: 'https://inference.canopywave.io/v1',
-    doc: 'https://canopywave.io/',
-    models: {
-      'minimax/minimax-m2.1': models['minimax-m2.1'],
-      'zai/glm-4.7': models['glm-4.7'],
-      'moonshotai/kimi-k2-thinking': models['kimi-k2-thinking'],
-      'deepseek/deepseek-chat-v3.2': models['deepseek-v3-2-exp'],
-    },
-    createModel: defaultModelCreator,
-  },
-  modelwatch: {
-    id: 'modelwatch',
-    source: 'built-in',
-    env: ['MODELWATCH_API_KEY'],
-    name: 'ModelWatch',
-    api: 'https://hub.modelwatch.dev/v1/',
-    doc: 'https://hub.modelwatch.dev/',
-    models: {
-      'qwen3-coder-plus': models['qwen3-coder-plus'],
-      'glm-4.7': models['glm-4.7'],
-      'gemini-2.5-flash': models['gemini-2.5-flash'],
-      'gemini-3-flash': models['gemini-3-flash-preview'],
-      'gemini-3-pro-preview': models['gemini-3-pro-preview'],
-      'claude-4-5-sonnet': models['claude-4-5-sonnet'],
-      'claude-haiku-4-5': models['claude-haiku-4-5'],
-      'claude-opus-4-5': models['claude-opus-4-5'],
-      'gpt-5.1': models['gpt-5.1'],
-      'gpt-5.1-codex-max': models['gpt-5.1-codex-max'],
-      'gpt-5.1-codex': models['gpt-5.1-codex'],
-      'gpt-5.1-codex-mini': models['gpt-5.1-codex-mini'],
-      'gpt-5.2': models['gpt-5.2'],
-      'gpt-5.2-codex': models['gpt-5.2-codex'],
-    },
-    createModel: (name, provider) => {
-      if (name.startsWith('claude-') || name.startsWith('gemini-')) {
-        return defaultAnthropicModelCreator(name, provider);
-      }
-      if (name.startsWith('gpt-')) {
-        return openaiModelResponseCreator(name, provider);
-      }
-      return defaultModelCreator(name, provider);
-    },
-  },
-};
 
 // value format: provider/model
 export type ModelAlias = Record<string, string>;
-export const modelAlias: ModelAlias = {
+const modelAlias: ModelAlias = {
   flash: 'google/gemini-2.5-flash',
   gemini: 'google/gemini-3-pro-preview',
   grok: 'xai/grok-4-1-fast',
@@ -2062,47 +43,96 @@ export const modelAlias: ModelAlias = {
 export type ModelInfo = {
   provider: Provider;
   model: Omit<Model, 'cost'>;
-  // m: LanguageModelV3;
   thinkingConfig?: Record<string, any>;
   _mCreator: () => Promise<LanguageModelV3>;
 };
 
-function mergeConfigProviders(
-  hookedProviders: ProvidersMap,
-  configProviders: Record<string, ProviderConfig>,
+/**
+ * Step 1: Get providers from plugin hooks
+ */
+async function getHookedProviders(context: Context): Promise<ProvidersMap> {
+  return context.apply({
+    hook: 'provider',
+    args: [
+      {
+        models,
+        createOpenAI,
+        createOpenAICompatible,
+        createAnthropic,
+      },
+    ],
+    memo: providers,
+    type: PluginHookType.SeriesLast,
+  });
+}
+
+/**
+ * Step 2: Merge two provider maps (config wins on overlap)
+ * Does NOT normalize - just merges
+ */
+function mergeProviders(
+  base: ProvidersMap,
+  override: Record<string, ProviderConfig>,
 ): ProvidersMap {
-  const mergedProviders = { ...hookedProviders };
-  Object.entries(configProviders).forEach(([providerId, config]) => {
-    let provider = mergedProviders[providerId] || {};
-    provider = defu(config, provider) as Provider;
-    if (!provider.createModel) {
-      const creatorMap = {
-        anthropic: defaultAnthropicModelCreator,
-        openai: defaultModelCreator,
-        responses: openaiModelResponseCreator,
-      };
-      const type = provider.apiFormat || 'openai';
-      provider.createModel = creatorMap[type];
+  const result = { ...base };
+  for (const [providerId, config] of Object.entries(override)) {
+    const existing = result[providerId] || {};
+    result[providerId] = defu(config, existing) as Provider;
+  }
+  return result;
+}
+
+/**
+ * Step 3: Normalize all providers
+ * Ensures every provider has: id, name, createModel, resolved models
+ */
+function normalizeProviders(providersMap: ProvidersMap): ProvidersMap {
+  const result: ProvidersMap = {};
+
+  for (const [providerId, provider] of Object.entries(providersMap)) {
+    const normalized = { ...provider } as Provider;
+
+    // Ensure id
+    if (!normalized.id) {
+      normalized.id = providerId;
     }
-    if (provider.models) {
-      for (const modelId in provider.models) {
-        const model = provider.models[modelId];
+
+    // Ensure name
+    if (!normalized.name) {
+      normalized.name = providerId;
+    }
+
+    // Resolve model string references
+    if (normalized.models) {
+      for (const modelId in normalized.models) {
+        const model = normalized.models[modelId];
+        let actualModel = {};
+        let extraInfo: Partial<Model> = {};
         if (typeof model === 'string') {
-          const actualModel = models[model];
-          assert(actualModel, `Model ${model} not exists.`);
-          provider.models[modelId] = actualModel;
+          actualModel = models[model.toLocaleLowerCase()];
+        } else {
+          const splitedModelId = modelId
+            .split('/')
+            .slice(-1)[0]
+            .toLocaleLowerCase();
+          actualModel = models[splitedModelId];
+          extraInfo = { ...model };
         }
+        assert(
+          actualModel,
+          `Model ${modelId} in provider ${providerId} not exists.`,
+        );
+        normalized.models[modelId] = {
+          ...actualModel,
+          ...extraInfo,
+        };
       }
     }
-    if (!provider.id) {
-      provider.id = providerId;
-    }
-    if (!provider.name) {
-      provider.name = providerId;
-    }
-    mergedProviders[providerId] = provider;
-  });
-  return mergedProviders;
+
+    result[providerId] = normalized;
+  }
+
+  return result;
 }
 
 /**
@@ -2113,11 +143,11 @@ function mergeConfigProviders(
  * @returns Updated providers map with global proxy applied
  */
 function applyGlobalProxyToProviders(
-  providers: ProvidersMap,
+  providersMap: ProvidersMap,
   globalHttpProxy: string,
 ): ProvidersMap {
   return Object.fromEntries(
-    Object.entries(providers).map(([id, prov]) => {
+    Object.entries(providersMap).map(([id, prov]) => {
       const provider = prov as Provider;
       // Skip if provider already has its own proxy
       if (provider.options?.httpProxy) {
@@ -2142,27 +172,18 @@ export async function resolveModelWithContext(
   name: string | null,
   context: Context,
 ) {
-  const hookedProviders = await context.apply({
-    hook: 'provider',
-    args: [
-      {
-        models,
-        defaultModelCreator,
-        createOpenAI,
-        createOpenAICompatible,
-        createAnthropic,
-      },
-    ],
-    memo: providers,
-    type: PluginHookType.SeriesLast,
-  });
+  // Step 1: Get hooked providers
+  const hookedProviders = await getHookedProviders(context);
 
-  let finalProviders = context.config.provider
-    ? mergeConfigProviders(hookedProviders, context.config.provider)
+  // Step 2: Merge with config providers (config wins)
+  const mergedProviders = context.config.provider
+    ? mergeProviders(hookedProviders, context.config.provider)
     : hookedProviders;
 
-  // Apply global proxy to ALL providers that don't have provider-level proxy
-  // This ensures both built-in and custom providers get the global proxy configuration
+  // Step 3: Normalize ALL providers
+  let finalProviders = normalizeProviders(mergedProviders);
+
+  // Step 4: Apply global proxy
   if (context.config.httpProxy) {
     finalProviders = applyGlobalProxyToProviders(
       finalProviders,
@@ -2216,22 +237,22 @@ export async function resolveModelWithContext(
   };
 }
 
-export async function resolveModel(
+async function resolveModel(
   name: string,
-  providers: ProvidersMap,
-  modelAlias: Record<string, string>,
+  providersMap: ProvidersMap,
+  modelAliasMap: Record<string, string>,
   globalConfigDir: string,
   setGlobalConfig: (key: string, value: string, isGlobal: boolean) => void,
 ): Promise<ModelInfo> {
-  const alias = modelAlias[name];
+  const alias = modelAliasMap[name];
   if (alias) {
     name = alias;
   }
   const [providerStr, ...modelNameArr] = name.split('/');
-  const provider = providers[providerStr];
+  const provider = providersMap[providerStr];
   assert(
     provider,
-    `Provider ${providerStr} not found, valid providers: ${Object.keys(providers).join(', ')}`,
+    `Provider ${providerStr} not found, valid providers: ${Object.keys(providersMap).join(', ')}`,
   );
   const modelId = modelNameArr.join('/');
   const model = provider.models[modelId] as Model;
@@ -2241,14 +262,12 @@ export async function resolveModel(
   );
   model.id = modelId;
   const mCreator = async () => {
-    let m: LanguageModelV3 | Promise<LanguageModelV3> = provider.createModel!(
-      modelId,
-      provider,
-      {
-        globalConfigDir,
-        setGlobalConfig,
-      },
-    );
+    let m: LanguageModelV3 | Promise<LanguageModelV3> = (
+      provider.createModel || createModelCreator
+    )(modelId, provider, {
+      globalConfigDir,
+      setGlobalConfig,
+    });
     if (isPromise(m)) {
       m = await m;
     }
