+++ src/modelMap.ts
@@ -1,0 +1,1047 @@
+import type { ModelMap } from './model';
+
+const baseModels: ModelMap = {
+  'deepseek-v3-0324': {
+    name: 'DeepSeek-V3-0324',
+    shortName: 'DeepSeek V3',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-06',
+    release_date: '2025-03-24',
+    last_updated: '2025-03-24',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 128000, output: 8192 },
+    aliases: ['deepseek-chat-v3-0324', 'deepseek-v3'],
+  },
+  'deepseek-v3-1': {
+    name: 'DeepSeek V3.1',
+    shortName: 'DeepSeek V3.1',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-07',
+    release_date: '2025-08-21',
+    last_updated: '2025-08-21',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 163840, output: 163840 },
+    aliases: ['deepseek-chat-v3.1', 'deepseek-v3.1', 'deepseek-v3-1-250821'],
+  },
+  'deepseek-v3-1-terminus': {
+    name: 'DeepSeek V3.1 Terminus',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-07',
+    release_date: '2025-09-22',
+    last_updated: '2025-09-22',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 65536 },
+    aliases: ['deepseek-v3.1-terminus'],
+  },
+  'deepseek-v3-2-exp': {
+    name: 'DeepSeek V3.2 Exp',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-09',
+    release_date: '2025-09-29',
+    last_updated: '2025-09-29',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 65536 },
+    aliases: ['deepseek-v3.2-exp'],
+  },
+  'deepseek-v3.2': {
+    name: 'DeepSeek V3.2',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-12',
+    release_date: '2025-12-01',
+    last_updated: '2025-12-01',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 65536 },
+    aliases: ['deepseek-chat', 'deepseek-chat-v3.2'],
+  },
+  'deepseek-v3.2-speciale': {
+    name: 'DeepSeek V3.2 Speciale',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: false,
+    knowledge: '2025-12',
+    release_date: '2025-12-01',
+    last_updated: '2025-12-01',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 131072 },
+  },
+  'deepseek-r1-0528': {
+    name: 'DeepSeek-R1-0528',
+    shortName: 'DeepSeek R1',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-06',
+    release_date: '2025-05-28',
+    last_updated: '2025-05-28',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 65536, output: 8192 },
+    aliases: ['deepseek-reasoner', 'deepseek-r1'],
+  },
+  'doubao-seed-1.6': {
+    name: 'Doubao Seed 1.6',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-01',
+    release_date: '2025-06-11',
+    last_updated: '2025-09-23',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 163840, output: 163840 },
+    aliases: ['doubao-seed-1-6-250615'],
+  },
+  'kimi-k2': {
+    name: 'Kimi K2',
+    attachment: false,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-10',
+    release_date: '2025-07-11',
+    last_updated: '2025-07-11',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 16384 },
+    aliases: ['kimi-k2-instruct'],
+  },
+  'kimi-k2-turbo-preview': {
+    name: 'Kimi K2 Turbo',
+    attachment: false,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-10',
+    release_date: '2025-07-14',
+    last_updated: '2025-07-14',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 16384 },
+  },
+  'kimi-k2-0905': {
+    name: 'Kimi K2 Instruct 0905',
+    shortName: 'Kimi K2 0905',
+    attachment: false,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-10',
+    release_date: '2025-09-05',
+    last_updated: '2025-09-05',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 262144, output: 16384 },
+    aliases: [
+      'kimi-k2-0905-preview',
+      'kimi-k2-instruct-0905',
+      'kimi-k2-250905',
+    ],
+  },
+  'kimi-k2-thinking': {
+    name: 'Kimi K2 Thinking',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-08',
+    release_date: '2025-11-06',
+    last_updated: '2025-11-06',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 262144, output: 262144 },
+  },
+  'kimi-k2-thinking-turbo': {
+    name: 'Kimi K2 Thinking Turbo',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-08',
+    release_date: '2025-11-06',
+    last_updated: '2025-11-06',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 262144, output: 262144 },
+  },
+  'kimi-k2.5': {
+    name: 'Kimi K2.5',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-01',
+    release_date: '2026-01-27',
+    last_updated: '2026-01-27',
+    modalities: { input: ['text', 'image', 'video'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 262144, output: 262144 },
+  },
+  'qwen3-coder-480b-a35b-instruct': {
+    name: 'Qwen3 Coder 480B A35B Instruct',
+    shortName: 'Qwen3 Coder',
+    attachment: false,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-04',
+    release_date: '2025-07-23',
+    last_updated: '2025-07-23',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 262144, output: 66536 },
+    aliases: ['qwen3-coder'],
+  },
+  'qwen3-coder-plus': {
+    name: 'Qwen3 Coder Plus',
+    attachment: false,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-04',
+    release_date: '2025-07-23',
+    last_updated: '2025-07-23',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 1048576, output: 65536 },
+  },
+  'qwen3-235b-a22b-07-25': {
+    name: 'Qwen3 235B A22B Instruct 2507',
+    shortName: 'Qwen3',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-04',
+    release_date: '2025-04-28',
+    last_updated: '2025-07-21',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 262144, output: 131072 },
+    aliases: ['qwen3-235b-a22b-instruct-2507'],
+  },
+  'qwen3-max': {
+    name: 'Qwen3 Max',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-09',
+    release_date: '2025-09-05',
+    last_updated: '2025-09-05',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 262144, output: 32768 },
+  },
+  'gemini-2.5-flash': {
+    name: 'Gemini 2.5 Flash',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-01',
+    release_date: '2025-03-20',
+    last_updated: '2025-06-05',
+    modalities: {
+      input: ['text', 'image', 'audio', 'video', 'pdf'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 1048576, output: 65536 },
+    aliases: ['gemini-2.5-flash-thinking'],
+  },
+  'gemini-2.5-flash-preview-09-2025': {
+    name: 'Gemini 2.5 Flash Preview 2025 09',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-01',
+    release_date: '2025-09-25',
+    last_updated: '2025-09-25',
+    modalities: {
+      input: ['text', 'image', 'audio', 'video', 'pdf'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 1048576, output: 65536 },
+  },
+  'gemini-2.5-flash-lite-preview-06-17': {
+    name: 'Gemini 2.5 Flash Lite Preview 06-17',
+    shortName: 'Gemini 2.5 Flash Lite',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-01',
+    release_date: '2025-06-17',
+    last_updated: '2025-06-17',
+    modalities: {
+      input: ['text', 'image', 'audio', 'video', 'pdf'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 65536, output: 65536 },
+    aliases: ['gemini-2.5-flash-lite'],
+  },
+  'gemini-2.5-pro': {
+    name: 'Gemini 2.5 Pro',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-01',
+    release_date: '2025-03-20',
+    last_updated: '2025-06-05',
+    modalities: {
+      input: ['text', 'image', 'audio', 'video', 'pdf'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 1048576, output: 65536 },
+  },
+  'gemini-3-pro-preview': {
+    name: 'Gemini 3 Pro Preview',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-01',
+    release_date: '2025-01-01',
+    last_updated: '2025-01-01',
+    modalities: {
+      input: ['text', 'image', 'audio', 'video', 'pdf'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 200000, output: 65536 },
+    aliases: ['gemini-3-pro', 'gemini-3-pro-low', 'gemini-3-pro-high'],
+  },
+  'gemini-3-flash-preview': {
+    name: 'Gemini 3 Flash Preview',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-01',
+    release_date: '2025-12-17',
+    last_updated: '2025-12-17',
+    modalities: {
+      input: ['text', 'image', 'video', 'audio', 'pdf'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 1048576, output: 65536 },
+    aliases: ['gemini-3-flash'],
+  },
+  'grok-4': {
+    name: 'Grok 4',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-07',
+    release_date: '2025-07-09',
+    last_updated: '2025-07-09',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 256000, output: 64000 },
+  },
+  'grok-code-fast-1': {
+    name: 'Grok Code Fast 1',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-08',
+    release_date: '2025-08-20',
+    last_updated: '2025-08-20',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 256000, output: 32000 },
+  },
+  'grok-4-fast': {
+    name: 'Grok 4 Fast',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-11',
+    release_date: '2025-08-19',
+    last_updated: '2025-08-19',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 2000000, output: 2000000 },
+  },
+  'grok-4.1-fast': {
+    name: 'Grok 4.1 Fast',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-10',
+    release_date: '2025-11-19',
+    last_updated: '2025-11-19',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 2000000, output: 2000000 },
+    aliases: ['grok-4-1-fast', 'grok-4-1-fast-non-reasoning'],
+  },
+  'claude-3-5-sonnet-20241022': {
+    name: 'Claude Sonnet 3.5 v2',
+    shortName: 'Sonnet 3.5',
+    attachment: true,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-04-30',
+    release_date: '2024-10-22',
+    last_updated: '2024-10-22',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 8192 },
+    aliases: ['claude-3.5-sonnet'],
+  },
+  'claude-3-7-sonnet': {
+    name: 'Claude Sonnet 3.7',
+    shortName: 'Sonnet 3.7',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-10-31',
+    release_date: '2025-02-19',
+    last_updated: '2025-02-19',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 64000 },
+    aliases: [
+      'claude-3.7-sonnet',
+      'claude-3.7-sonnet-thought',
+      'claude-3-7-sonnet-20250219',
+      'claude-3-7-sonnet-20250219-thinking',
+    ],
+  },
+  'claude-4-sonnet': {
+    name: 'Claude Sonnet 4',
+    shortName: 'Sonnet 4',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-03-31',
+    release_date: '2025-05-22',
+    last_updated: '2025-05-22',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 64000 },
+    aliases: ['claude-sonnet-4', 'claude-sonnet-4-20250514'],
+  },
+  'claude-4-opus': {
+    name: 'Claude Opus 4',
+    shortName: 'Opus 4',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-03-31',
+    release_date: '2025-05-22',
+    last_updated: '2025-05-22',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 32000 },
+    aliases: ['claude-opus-4', 'claude-opus-4-20250514'],
+  },
+  'gpt-oss-120b': {
+    name: 'GPT OSS 120B',
+    shortName: 'GPT OSS',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-08',
+    release_date: '2025-08-05',
+    last_updated: '2025-08-05',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 32768 },
+    aliases: ['gpt-oss-120b-medium'],
+  },
+  'gpt-5.1-codex': {
+    name: 'GPT-5.1-Codex',
+    attachment: false,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-09-30',
+    release_date: '2025-11-13',
+    last_updated: '2025-11-13',
+    modalities: {
+      input: ['text', 'image'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 400000, output: 128000 },
+  },
+  'gpt-5.1-codex-mini': {
+    name: 'GPT-5.1-Codex-mini',
+    attachment: false,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-09-30',
+    release_date: '2025-11-13',
+    last_updated: '2025-11-13',
+    modalities: {
+      input: ['text', 'image'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 400000, output: 100000 },
+  },
+  'gpt-5.1-codex-max': {
+    name: 'GPT 5.1 Codex Max',
+    attachment: true,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-09-30',
+    release_date: '2025-11-13',
+    last_updated: '2025-11-13',
+    modalities: {
+      input: ['text', 'image'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 400000, output: 128000 },
+  },
+  'gpt-5.1': {
+    name: 'GPT-5.1',
+    attachment: true,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-09-30',
+    release_date: '2025-11-13',
+    last_updated: '2025-11-13',
+    modalities: {
+      input: ['text', 'image'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 400000, output: 128000 },
+  },
+  'gpt-5': {
+    name: 'GPT-5',
+    attachment: true,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-09-30',
+    release_date: '2025-08-07',
+    last_updated: '2025-08-07',
+    modalities: {
+      input: ['text', 'audio', 'image', 'video'],
+      output: ['text', 'audio', 'image'],
+    },
+    open_weights: false,
+    limit: { context: 400000, output: 128000 },
+  },
+  'gpt-5-mini': {
+    name: 'GPT-5 Mini',
+    attachment: true,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-05-30',
+    release_date: '2025-08-07',
+    last_updated: '2025-08-07',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 272000, output: 128000 },
+  },
+  'gpt-5-codex': {
+    name: 'GPT-5-Codex',
+    attachment: false,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-09-30',
+    release_date: '2025-09-15',
+    last_updated: '2025-09-15',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 128000, output: 64000 },
+  },
+  'gpt-5.2': {
+    name: 'GPT-5.2',
+    attachment: true,
+    reasoning: true,
+    temperature: false,
+    knowledge: '2025-08-31',
+    tool_call: true,
+    release_date: '2025-12-11',
+    last_updated: '2025-12-11',
+    modalities: { input: ['text', 'image'], output: ['text', 'image'] },
+    open_weights: false,
+    limit: { context: 400000, output: 128000 },
+  },
+  'gpt-5.2-pro': {
+    name: 'GPT-5.2 Pro',
+    attachment: true,
+    reasoning: true,
+    temperature: false,
+    knowledge: '2025-08-31',
+    tool_call: true,
+    release_date: '2025-12-11',
+    last_updated: '2025-12-11',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 400000, output: 128000 },
+  },
+  'gpt-5.2-codex': {
+    name: 'GPT-5.2-Codex',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-08-31',
+    release_date: '2026-01-14',
+    last_updated: '2026-01-14',
+    modalities: {
+      input: ['text', 'image'],
+      output: ['text'],
+    },
+    open_weights: false,
+    limit: { context: 400000, output: 128000 },
+  },
+  'gpt-4.1': {
+    name: 'GPT-4.1',
+    attachment: true,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2024-04',
+    release_date: '2025-04-14',
+    last_updated: '2025-04-14',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 1047576, output: 32768 },
+  },
+  'gpt-4': {
+    name: 'GPT-4',
+    attachment: true,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2023-11',
+    release_date: '2023-11-06',
+    last_updated: '2024-04-09',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 8192, output: 8192 },
+  },
+  'gpt-4o': {
+    name: 'GPT-4o',
+    attachment: true,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2023-09',
+    release_date: '2024-05-13',
+    last_updated: '2024-05-13',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 128000, output: 16384 },
+  },
+  o3: {
+    name: 'o3',
+    attachment: true,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-05',
+    release_date: '2025-04-16',
+    last_updated: '2025-04-16',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 100000 },
+  },
+  'o3-pro': {
+    name: 'o3-pro',
+    attachment: true,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-05',
+    release_date: '2025-06-10',
+    last_updated: '2025-06-10',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 100000 },
+  },
+  'o3-mini': {
+    name: 'o3-mini',
+    attachment: false,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-05',
+    release_date: '2024-12-20',
+    last_updated: '2025-01-29',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 100000 },
+  },
+  'o4-mini': {
+    name: 'o4-mini',
+    attachment: true,
+    reasoning: true,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-05',
+    release_date: '2025-04-16',
+    last_updated: '2025-04-16',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 100000 },
+  },
+  'glm-4.5': {
+    name: 'GLM-4.5',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-04',
+    release_date: '2025-07-28',
+    last_updated: '2025-07-28',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 98304 },
+    aliases: [],
+  },
+  'glm-4.5-air': {
+    name: 'GLM-4.5-Air',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-04',
+    release_date: '2025-07-28',
+    last_updated: '2025-07-28',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 98304 },
+  },
+  'glm-4.5-flash': {
+    name: 'GLM-4.5-Flash',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-04',
+    release_date: '2025-07-28',
+    last_updated: '2025-07-28',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 98304 },
+  },
+  'glm-4.5v': {
+    name: 'GLM-4.5V',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-04',
+    release_date: '2025-08-11',
+    last_updated: '2025-08-11',
+    modalities: { input: ['text', 'image', 'video'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 64000, output: 16384 },
+  },
+  'glm-4.6': {
+    name: 'GLM-4.6',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-04',
+    release_date: '2025-09-30',
+    last_updated: '2025-09-30',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 204800, output: 131072 },
+  },
+  'glm-4.6v': {
+    name: 'GLM-4.6V',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-04',
+    release_date: '2025-12-08',
+    last_updated: '2025-12-08',
+    modalities: { input: ['text', 'image', 'video'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 24000 },
+  },
+  'glm-4.7': {
+    name: 'GLM-4.7',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-04',
+    release_date: '2025-12-22',
+    last_updated: '2025-12-22',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 204800, output: 131072 },
+    aliases: ['zai-glm-4.7', 'glm4.7'],
+  },
+  'sonoma-dusk-alpha': {
+    name: 'Sonoma Dusk Alpha',
+    attachment: true,
+    reasoning: false,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-09',
+    release_date: '2024-09-05',
+    last_updated: '2024-09-05',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 2000000, output: 2000000 },
+  },
+  'sonoma-sky-alpha': {
+    name: 'Sonoma Sky Alpha',
+    attachment: true,
+    reasoning: false,
+    temperature: false,
+    tool_call: true,
+    knowledge: '2024-09',
+    release_date: '2024-09-05',
+    last_updated: '2024-09-05',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 2000000, output: 2000000 },
+  },
+  'claude-4.1-opus': {
+    name: 'Claude Opus 4.1',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-03-31',
+    release_date: '2025-08-05',
+    last_updated: '2025-08-05',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 32000 },
+    aliases: [
+      'claude-opus-41',
+      'claude-opus-4-1-20250805',
+      'claude-opus-4-1',
+      'claude-opus-4.1',
+    ],
+  },
+  'claude-4-5-sonnet': {
+    name: 'Claude Sonnet 4.5 (Preview)',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-03-31',
+    release_date: '2025-09-29',
+    last_updated: '2025-09-29',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 32000 },
+    aliases: [
+      'claude-sonnet-4.5',
+      'claude-sonnet-4-5-20250929',
+      'claude-sonnet-4-5',
+      'claude-sonnet-4-5-thinking',
+    ],
+  },
+  'claude-haiku-4-5': {
+    name: 'Claude Haiku 4.5',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-02-31',
+    release_date: '2025-10-15',
+    last_updated: '2025-10-15',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 64000 },
+    aliases: ['claude-haiku-4.5'],
+  },
+  'claude-opus-4-5': {
+    name: 'Claude Opus 4.5',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-03-31',
+    release_date: '2025-11-24',
+    last_updated: '2025-11-24',
+    modalities: { input: ['text', 'image'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 200000, output: 64000 },
+    aliases: ['claude-opus-4.5', 'claude-opus-4-5-thinking'],
+  },
+  'ling-1t': {
+    name: 'InclusionAI Ling-1T',
+    attachment: true,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-10-09',
+    release_date: '2025-10-09',
+    last_updated: '2025-10-09',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 128000, output: 32000 },
+  },
+  'ring-1t': {
+    name: 'InclusionAI Ring-1T',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-10-14',
+    release_date: '2025-10-14',
+    last_updated: '2025-10-14',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 128000, output: 32000 },
+  },
+  'ring-flash-2.0': {
+    name: 'InclusionAI Ring-flash-2.0',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-09-09',
+    release_date: '2025-09-15',
+    last_updated: '2025-09-16',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 128000, output: 32000 },
+  },
+  'ling-flash-2.0': {
+    name: 'InclusionAI Ling-flash-2.0',
+    attachment: true,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-09-09',
+    release_date: '2025-09-15',
+    last_updated: '2025-09-16',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 128000, output: 32000 },
+  },
+  'ring-mini-2.0': {
+    name: 'InclusionAI Ring-mini-2.0',
+    attachment: true,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-09-09',
+    release_date: '2025-09-15',
+    last_updated: '2025-09-16',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 128000, output: 32000 },
+  },
+  'ling-mini-2.0': {
+    name: 'InclusionAI Ling-mini-2.0',
+    attachment: true,
+    reasoning: false,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-09-09',
+    release_date: '2025-09-15',
+    last_updated: '2025-09-16',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: false,
+    limit: { context: 128000, output: 32000 },
+  },
+  'minimax-m2': {
+    name: 'Minimax-M2',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '',
+    release_date: '2025-10-27',
+    last_updated: '2025-10-27',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 196608, output: 64000 },
+  },
+  'mimo-v2-flash': {
+    name: 'MiMo V2 Flash',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '2025-12',
+    release_date: '2025-12-17',
+    last_updated: '2025-12-17',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 131072, output: 32768 },
+  },
+  'minimax-m2.1': {
+    name: 'MiniMax-M2.1',
+    attachment: false,
+    reasoning: true,
+    temperature: true,
+    tool_call: true,
+    knowledge: '',
+    release_date: '2025-12-23',
+    last_updated: '2025-12-23',
+    modalities: { input: ['text'], output: ['text'] },
+    open_weights: true,
+    limit: { context: 204800, output: 131072 },
+  },
+};
+
+export const models = (() => {
+  const models = { ...baseModels };
+  for (const modelId in models) {
+    const model = models[modelId];
+    if (model.aliases) {
+      for (const alias of model.aliases) {
+        models[alias] = model;
+      }
+    }
+  }
+  return models;
+})();
