+++ src/providers/aihubmix.ts
@@ -1,0 +1,34 @@
+import { type Provider } from './types';
+
+export const aihubmixProvider: Provider = {
+  id: 'aihubmix',
+  source: 'built-in',
+  env: ['AIHUBMIX_API_KEY'],
+  name: 'AIHubMix',
+  api: 'https://aihubmix.com/v1',
+  doc: 'https://docs.aihubmix.com/',
+  models: {
+    'gemini-2.5-pro': {},
+    'gemini-2.5-flash': {},
+    'gemini-2.5-flash-lite': {},
+    'DeepSeek-R1': {},
+    'DeepSeek-V3': {},
+    'claude-opus-4-20250514': {},
+    'claude-opus-4-1': {},
+    'claude-sonnet-4-20250514': {},
+    'claude-sonnet-4-5': {},
+    'claude-3-7-sonnet-20250219': {},
+    'claude-3-5-sonnet-20241022': {},
+    'gpt-4.1': {},
+    'gpt-4': {},
+    'gpt-4o': {},
+    o3: {},
+    'o3-mini': {},
+    'o4-mini': {},
+    'gpt-5': {},
+    'gpt-5-mini': {},
+    'glm-4.6': {},
+    'kimi-k2-thinking': {},
+    'kimi-k2-turbo-preview': {},
+  },
+};
